윷놀이 (#63)
* feat: room 목록 페이지 렌더링 시 서버와 통신하도록 axios 활용

* feat: 방 리스트 로고 사진 변경 (상대경로 이용)

* feat: 윷놀이 보드 구현

* feat: 윷놀이판(보드) 프론트 구현

* feat: Point 생성 및 관계 초기화 구현

* feat: Yut 구현

* feat: RandomYutThrowStrategy 구현

* feat: piece 객체 구현

* feat: Board객체 수정 및 Piece들 잡기&업기를 포함한 움직임 로직 구현

* refactor: clean up codes

* test: PieceTest 테스트 로직 변경

* feat: moveVerifier 구현

* test: YutThrowTest 추가

* test: test 케이스 추가

* feat: 윷놀이게임 구현

* feat: 말을 움직이지 않고 연속으로 윳을 던지려고 했을 때 예외처리

* test: TurnTest 구현

* feat: turn 구현 및 YutNoriGame과 연결

* refactor: change package structure

* feat: 다음 턴 관리 구현, 보드 생성 방법 변경

* refactor: 사용하지 않는 메소드 제거

* refactor: YurnoriBaseTest 생성

* test: 윷놀이게임 test case 추가

* test: 테스트에 빽도 추가

* refactor: YUTNORI GameKind추가

* feat: 모든 말이 끝난 플레이어가 턴에서 제외되도록 변경

* refactor: 패키지 정리

* refactor: clean up codes

* refactor: 패키지 구조 변경

* feat: setup for WebSocketAcceptanceTest

* feat: user MockController to put User in session

* feat: 윷던지기 구현

* refactor: 피드백 반영

* feat: 로그인 구현 로직 변경 및 말 움직이기 구현

* refactor: Route PointName 갖게 바꿈

* refactor: implement MoveResultDtos

* refactor: add MoveResultDtos

* refactor: final 키워드 추가

* feat: yutnori 프론트

* add: 로그인 변경 부분 반영

* feat: 윷놀이 게임 시작까지 가능

* feat: YutnoriGameService 구현

* fix: fix all test bugs

* refactor: create dtos in service layer

* feat: 게임 끝나는 조건 구현

* feat: 게임 초기화(시작)은 RoomApiController를 거치도록 변경
- YutnoriGameController에 게임 초기화하던 메서드 제거
- Dto 구조 변경

* feat: YutThrowStrategy 빈 등록하여 테스트에서는 도만 던져지도록 적용

* feat: Turn 관련 에러 상황 구체적으로 표현하는 exception 구현

* refactor: Yuts 불변 객체 적용

* feat: 던진 yut 목록 전달 필요해서 response 구조 변경
- throw yut, move piece 요청 모두 던진 yut 목록 포함하도록

* feat: 프론트 기본 게임 동작 구현

* refactor: 코드 컨벤션 및 빠트린 커밋(YutnoriGameService)

* fix: bug fix (보드 초기화 잘못되고 있었음)

* refactor: 불필요한 파일 제거 및 코드 포맷팅

* feat: 차례에 따라 말 움직이기, 윷 던지기 프론트에서 버튼 활성화/비활성화

* feat: 해당 움직임으로 승리했는지(해당 유저 게임 끝) 확인하도록 구현
- 모두에게 '성공' alert

* refactor: 라우팅 경로 변경

* feat: 윷놀이 프론트 기능 추가
- 말 클릭시 표시
- 윷 클릭시 표시

* feat: 게임 종류별 리스팅 및 방생성 처리

* refactor: 테스트코드 리팩터링 (YutnoriGameControllerTest, MockController)

* test: 말 두번 움직이기 테스트 추가

* feat: 도개걸윷모 확률 로직 변경

* feat: 움직일 수 없는 경우 체크 (해당 포인트에 말이 없는 경우)

+++ src/main/java/com/olbimacoojam/heaven/draw/application/DrawService.java
@@ -22,20 +22,20 @@ public class DrawService {
         this.userService = userService;
     }
 
-    public DrawResponse updateGame(long userId, long roomId, DrawCreateRequest drawCreateRequest) {
+    public DrawResponse updateGame(long userId, int roomId, DrawCreateRequest drawCreateRequest) {
         Draw draw = getDraw(userId, roomId);
         draw.startGame(drawCreateRequest.getPersonCount(), drawCreateRequest.getWhackCount());
 
         return new DrawResponse(draw.getLots());
     }
 
-    public DrawResponse initGame(long userId, long roomId) {
+    public DrawResponse initGame(long userId, int roomId) {
         Draw draw = getDraw(userId, roomId);
 
         return new DrawResponse(draw.getLots());
     }
 
-    private Draw getDraw(long userId, long roomId) {
+    private Draw getDraw(long userId, int roomId) {
         Room room = roomService.findById(roomId);
         Draw draw = (Draw) room.getGame();
         User user = userService.findById(userId);
