+++ src/test/java/com/olbimacoojam/heaven/draw/application/DrawServiceTest.java
@@ -39,15 +39,15 @@ class DrawServiceTest {
     @Test
     void updateGame() {
         Game game = new Draw();
-        Room room = new Room(1L, game);
+        Room room = new Room(1, game);
         User user = new User(1L, "martin", "token");
         Draw draw = (Draw)room.getGame();
 
-        given(roomService.findById(any(Long.class))).willReturn(room);
+        given(roomService.findById(any(Integer.class))).willReturn(room);
         given(mockRoom.getGame()).willReturn(game);
         given(userService.findById(any(Long.class))).willReturn(user);
 
-        DrawResponse drawResponse = drawService.updateGame(1L, 1L, new DrawCreateRequest(10, 1));
+        DrawResponse drawResponse = drawService.updateGame(1L, 1, new DrawCreateRequest(10, 1));
 
         assertThat(drawResponse.getLots().getLots().stream().filter(Lot.PASS::equals).count()).isEqualTo(9L);
         assertThat(drawResponse.getLots().getLots().stream().filter(Lot.WHACK::equals).count()).isEqualTo(1L);
@@ -56,15 +56,15 @@ class DrawServiceTest {
     @Test
     void initGame() {
         Game game = new Draw();
-        Room room = new Room(1L, game);
+        Room room = new Room(1, game);
         User user = new User(1L, "martin", "token");
         Draw draw = (Draw)room.getGame();
 
-        given(roomService.findById(any(Long.class))).willReturn(room);
+        given(roomService.findById(any(Integer.class))).willReturn(room);
         given(mockRoom.getGame()).willReturn(game);
         given(userService.findById(any(Long.class))).willReturn(user);
 
-        DrawResponse drawResponse = drawService.initGame(1L, 1L);
+        DrawResponse drawResponse = drawService.initGame(1L, 1);
 
         assertThat(drawResponse.getLots().getLots().stream().filter(Lot.PASS::equals).count()).isEqualTo(3L);
         assertThat(drawResponse.getLots().getLots().stream().filter(Lot.WHACK::equals).count()).isEqualTo(1L);
